refactor: s3Client 분리
