+++ src/main/java/kr/co/conceptbe/image/application/ImageService.java
@@ -1,20 +1,12 @@
 package kr.co.conceptbe.image.application;
 
-import com.amazonaws.services.s3.AmazonS3;
-import com.amazonaws.services.s3.model.DeleteObjectRequest;
-import com.amazonaws.services.s3.model.PutObjectRequest;
-import java.io.File;
-import java.io.IOException;
 import java.util.List;
-import java.util.Objects;
 import kr.co.conceptbe.auth.presentation.dto.AuthCredentials;
 import kr.co.conceptbe.image.domain.IdeaValidator;
 import kr.co.conceptbe.image.domain.Image;
 import kr.co.conceptbe.image.domain.ImageChecker;
 import kr.co.conceptbe.image.domain.ImageRepository;
-import kr.co.conceptbe.image.domain.UploadFile;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
@@ -24,12 +16,10 @@ import org.springframework.web.multipart.MultipartFile;
 @RequiredArgsConstructor
 public class ImageService {
 
-    @Value("${s3.bucket}")
-    private String bucket;
-    private final AmazonS3 amazonS3;
     private final ImageRepository imageRepository;
     private final IdeaValidator ideaValidator;
     private final ImageChecker imageChecker;
+    private final S3Client s3Client;
 
     public void save(Long ideaId, AuthCredentials authCredentials, List<MultipartFile> files) {
         ideaValidator.validateIdea(ideaId, authCredentials.id());
@@ -39,43 +29,9 @@ public class ImageService {
 
     private void uploadImages(Long ideaId, List<MultipartFile> files) {
         files.stream()
-            .map(this::upload)
+            .map(s3Client::upload)
             .map(imageUrl -> new Image(ideaId, imageUrl))
             .forEach(imageRepository::save);
-    }
-
-    private String upload(MultipartFile multipartFile) {
-        try {
-            UploadFile uploadFile = UploadFile.from(multipartFile);
-            fileUpload(uploadFile);
-            return uploadFile.getOriginalFilename();
-        } catch (IOException exception) {
-            throw new RuntimeException(exception);
-        }
-    }
-
-    private void fileUpload(MultipartFile multipartFile) throws IOException {
-        File tempFile = null;
-
-        try {
-            tempFile = File.createTempFile("upload_", ".tmp");
-            multipartFile.transferTo(tempFile);
-            amazonS3.putObject(new PutObjectRequest(
-                bucket,
-                multipartFile.getOriginalFilename(),
-                tempFile
-            ));
-        } catch (IOException exception) {
-            throw new IOException(exception);
-        } finally {
-            removeTempFileIfExists(tempFile);
-        }
-    }
-
-    private void removeTempFileIfExists(File tempFile) {
-        if (Objects.nonNull(tempFile) && tempFile.exists()) {
-            tempFile.delete();
-        }
     }
 
     public void update(
@@ -87,7 +43,7 @@ public class ImageService {
         ideaValidator.validateIdea(ideaId, authCredentials.id());
         List<Image> imagesToDeleted = getImagesToDeleted(ideaId, imageIds, additionFiles.size());
         imagesToDeleted.forEach(this::deleteImage);
-        additionFiles.forEach(this::upload);
+        additionFiles.forEach(s3Client::upload);
     }
 
     private List<Image> getImagesToDeleted(
@@ -117,7 +73,7 @@ public class ImageService {
     }
 
     private void deleteImage(Image image) {
-        amazonS3.deleteObject(new DeleteObjectRequest(bucket, image.getImageUrl()));
+        s3Client.delete(image.getImageUrl());
         imageRepository.delete(image);
     }
